+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -14,6 +14,8 @@
     ([#26009](https://github.com/Azure/azure-sdk-for-python/pull/26009))
 - Implement instrumentation statsbeat
     ([#26023](https://github.com/Azure/azure-sdk-for-python/pull/26023))
+- Implement statsbeat shutdown
+    ([#26023](https://github.com/Azure/azure-sdk-for-python/pull/26023))
 
 ### Breaking Changes